% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ame_forest_group_plot}
\alias{ame_forest_group_plot}
\title{Combined all outcomes on a single forest plot}
\usage{
ame_forest_group_plot(
  ame_data,
  title = "",
  save = FALSE,
  filename = "ame_forest_group_plot.png"
)
}
\description{
Combined all outcomes on a single forest plot
}
