% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_effect_on_treatment.R
\name{average_effect_on_treatment}
\alias{average_effect_on_treatment}
\title{Average Treatment Effect on Treatment For All Variables}
\usage{
average_effect_on_treatment(fit, data, save = FALSE)
}
\description{
Average Treatment Effect on Treatment For All Variables
}
