% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_marginal_effect.R
\name{average_marginal_effect}
\alias{average_marginal_effect}
\title{Average Marginal Effect For All Variables}
\usage{
average_marginal_effect(fit, data, save = FALSE)
}
\description{
Average Marginal Effect For All Variables
}
