% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_effect_on_treatment.R
\name{cal_att_base}
\alias{cal_att_base}
\title{Calculate Average Treatment Effect on Treatment for Frequentist}
\usage{
cal_att_base(fit, data, var)
}
\description{
Calculate Average Treatment Effect on Treatment for Frequentist
}
