% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_marginal_effect.R
\name{calc_ame_bayesian}
\alias{calc_ame_bayesian}
\title{Calculate Bayesian AMEs for a single variable}
\usage{
calc_ame_bayesian(fit, poststrat_data, var_name, baseline_draws, ndraws = NULL)
}
\arguments{
\item{fit}{A fitted stanreg or brmsfit model.}

\item{poststrat_data}{The main post-stratification data frame.}

\item{var_name}{The name of the variable (string) to calculate AMEs for.}

\item{baseline_draws}{The vector of posterior draws for the grand mean.}
}
\value{
A data frame of AME posterior draws, or NULL if var_name is not a factor.
}
\description{
Calculate Bayesian AMEs for a single variable
}
