% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_effect_on_treatment.R
\name{calc_att_stan}
\alias{calc_att_stan}
\title{Calculate Average Treatment Effect on Treatment for Stan

this is essentially just a change in data input
i.e. subpopulation profile
relative to the ame calculation for the whole population}
\usage{
calc_att_stan(fit, data, var)
}
\description{
Calculate Average Treatment Effect on Treatment for Stan

this is essentially just a change in data input
i.e. subpopulation profile
relative to the ame calculation for the whole population
}
