% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lfs_synth_data.R
\name{clean_lfs_national_data}
\alias{clean_lfs_national_data}
\title{clean labour force survey data for all England}
\usage{
clean_lfs_national_data()
}
\description{
clean labour force survey data for all England
}
