% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_treatment_effect.R
\name{cond_treatment_effect}
\alias{cond_treatment_effect}
\title{Conditional Average Treatment Effect (CATE)}
\usage{
cond_treatment_effect(fit, survey_dat, mrp_dat, interaction = "workingstatus")
}
\arguments{
\item{fit}{glm or stan model output}

\item{survey_dat}{data frame}

\item{interaction}{Name of the interaction term}
}
\value{
List
}
\description{
Conditional Average Treatment Effect (CATE)
}
