% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_poststrat_inputs.R
\name{create_target_pop_data}
\alias{create_target_pop_data}
\title{Create target population data

Use individual level Resident survey data
so can estimate full joint distribution}
\usage{
create_target_pop_data(covariate_data, save = FALSE)
}
\arguments{
\item{covariate_data}{Covariate data}
}
\value{
dataframe of levels and joint probability
}
\description{
Create target population data

Use individual level Resident survey data
so can estimate full joint distribution
}
