% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_model}
\alias{fit_model}
\title{Fit a single model}
\usage{
fit_model(dat, outcome_var, formula_rhs, model_fun, common_args, ...)
}
\arguments{
\item{dat}{The data to use for modelling (either a data.frame or a mids object).}

\item{outcome_var}{A string, the name of the outcome variable.}

\item{formula_rhs}{A string, the right-hand side of the model formula.}

\item{model_fun}{Modelling function to call (e.g., stan_glmer).
Ignored if 'dat' is a 'mids' object.}

\item{common_args}{A list of common arguments for the model function.}

\item{...}{Additional arguments passed to the modelling function.}
}
\value{
A fitted model object.
}
\description{
Helper function to fit one model. It detects if 'dat' is a 'mids' object
and calls brms::brm_multiple, or if 'dat' is a data.frame and calls
the specified 'model_fun'.
}
