% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_models}
\alias{fit_models}
\title{Fit health literacy, numeracy and ICT models}
\usage{
fit_models(survey_data, stan = TRUE, save = FALSE, ...)
}
\arguments{
\item{survey_data}{List of data frames containing the cleaned survey data}

\item{stan}{Logical indicating whether to use Stan or not}

\item{save}{Logical indicating whether to save the fitted models}

\item{...}{Additional arguments to pass to the Stan model}
}
\value{
List of fitted models
}
\description{
Fit health literacy, numeracy and ICT models
}
