% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_england_census_props}
\alias{get_england_census_props}
\title{Proportion marginal tables other sources
from 2021 Census (see nomis)}
\usage{
get_england_census_props()
}
\description{
Proportion marginal tables other sources
from 2021 Census (see nomis)
}
