% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_newham_census_props}
\alias{get_newham_census_props}
\title{Proportion marginal tables other sources
from Census}
\usage{
get_newham_census_props(equivalise_income = FALSE)
}
\description{
Proportion marginal tables other sources
from Census
}
