% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_imputed_domain_par.R
\name{prepare_imputed_domain}
\alias{prepare_imputed_domain}
\title{Serial Version of Prepares a combined and imputed 'mids' object for a specific skill domain.}
\usage{
prepare_imputed_domain(
  domain_name,
  piaac_data,
  sfl_data,
  all_covariates,
  m_imputations,
  n_maxit = 20
)
}
\arguments{
\item{domain_name}{The name of the skill domain (e.g., "lit", "num") as a string.}

\item{piaac_data}{The cleaned PIAAC_survey_data (a list of data frames).}

\item{sfl_data}{The cleaned SfL2011_survey_data (a list of data frames).}

\item{all_covariates}{A character vector of all covariate names.}

\item{m_imputations}{The number of imputations.}
}
\value{
A 'mids' object.
}
\description{
Serial Version of Prepares a combined and imputed 'mids' object for a specific skill domain.
}
