% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_imputed_domain_par.R
\name{prepare_imputed_domain_par}
\alias{prepare_imputed_domain_par}
\title{Prepares a combined and imputed 'mids' object for a specific skill domain}
\usage{
prepare_imputed_domain_par(
  domain_name,
  piaac_data,
  sfl_data,
  all_covariates,
  m_imputations,
  n_maxit = 20,
  noise_sd = 0,
  include_outcome = FALSE
)
}
\arguments{
\item{domain_name}{The name of the skill domain (e.g., "lit", "num") as a string.}

\item{piaac_data}{The cleaned PIAAC_survey_data (a list of data frames).}

\item{sfl_data}{The cleaned SfL2011_survey_data (a list of data frames).}

\item{all_covariates}{A character vector of all covariate names.}

\item{m_imputations}{The number of imputations (e.g., 5).}

\item{noise_sd}{Proxy noise standard deviation}
}
\value{
A 'mids' object from mice package
}
\description{
Prepares a combined and imputed 'mids' object for a specific skill domain
}
