% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rank_group_plot}
\alias{rank_group_plot}
\title{all outcomes on a single rank stacked bar plot}
\usage{
rank_group_plot(
  ame_data,
  max_rank = 3,
  title = "",
  abs_val = TRUE,
  save = FALSE
)
}
\arguments{
\item{abs}{logical, if TRUE, use marginal effect absolute values}
}
\description{
all outcomes on a single rank stacked bar plot
}
