% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rank_plot_by_var}
\alias{rank_plot_by_var}
\title{stacked bar chart with variable on the x-axis}
\usage{
rank_plot_by_var(ps_var, max_rank = 5, title = "", save = FALSE)
}
\description{
stacked bar chart with variable on the x-axis
}
