% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{res_survey_prop_tables}
\alias{res_survey_prop_tables}
\title{Proportion marginal tables from resident survey
report summary tables (weighted)}
\usage{
res_survey_prop_tables()
}
\description{
Proportion marginal tables from resident survey
report summary tables (weighted)
}
