% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{sucra_table}
\alias{sucra_table}
\title{Ranks summary table

Includes cumulative rank, SUCRA and expected rank}
\usage{
sucra_table(ame_data, max_rank = 3, threshold = 0.2, abs_val = TRUE)
}
\description{
Ranks summary table

Includes cumulative rank, SUCRA and expected rank
}
