% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_data}
\alias{clean_data}
\title{Clean survey data}
\usage{
clean_data(data)
}
\arguments{
\item{data}{Survey data at individual level i.e. Skills for Life data}
}
\value{
List of cleaned data sets
}
\description{
Clean survey data
}
