% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lfs_synth_data.R
\name{create_lfs_synth_data}
\alias{create_lfs_synth_data}
\title{Create labour force survey synthetic data}
\usage{
create_lfs_synth_data(target_marginals_props = NULL)
}
\arguments{
\item{target_marginals_props}{Target marginal proportions as a list.}
}
\description{
Using {simPop} package to perform iterative proportional fitting (IPF).
}
\examples{
# Verification: Check Marginals
\dontrun{
result_list <-
  lapply(vars_to_simulate,
         \(x) {
           prop_table <- prop.table(table(synthetic_data_final[[x]]))
           df_output <- as.data.frame(prop_table, stringsAsFactors = FALSE)

           colnames(df_output) <- names(target_marginals_props[[x]])
           return(df_output)
         }) |>
  setNames(vars_to_simulate)
}
}
