% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_poststrat_inputs.R
\name{create_NRS_prob_data}
\alias{create_NRS_prob_data}
\title{Create Newham Resident Survey probability data}
\usage{
create_NRS_prob_data(save_output = TRUE)
}
\description{
from resident survey individual level data
joint distribution
}
