% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_poststrat_inputs.R
\name{create_target_pop_data}
\alias{create_target_pop_data}
\title{Calculate product of probabilities, assuming independence}
\usage{
create_target_pop_data(covariate_prob_data, save_output = FALSE)
}
\arguments{
\item{save_output}{logical}

\item{additional_prob_data}{non-NRS data}

\item{covariate_data}{Covariate data from SfL}
}
\value{
dataframe of levels and joint probability
}
\description{
Calculate product of probabilities, assuming independence
}
