% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_marginal_effect.R
\name{strat_marginal_effect}
\alias{strat_marginal_effect}
\title{Stratified Average Marginal Effect}
\usage{
strat_marginal_effect(fit, data)
}
\description{
Stratified Average Marginal Effect
}
