% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{sucra_table}
\alias{sucra_table}
\title{Ranks SUCRA summary table

Includes cumulative rank, SUCRA and expected rank}
\usage{
sucra_table(ame_data, max_rank = NA, threshold = 0, abs_val = TRUE)
}
\arguments{
\item{ame_data}{}

\item{max_rank}{An integer specifying the maximum rank to display on the x-axis.
Defaults to `NA` for all ranks.}

\item{threshold}{A numeric value between 0 and 1. Variables that never achieve a
cumulative rank value above this threshold will be filtered out to declutter the plot.
Defaults to `0`.}

\item{abs_val}{A logical value. If `TRUE`, ranking is based on the absolute magnitude
of `ame_base` (largest absolute value gets the best rank). Defaults to `FALSE`.}
}
\description{
Ranks SUCRA summary table

Includes cumulative rank, SUCRA and expected rank
}
